Add distance update functionality and optimize data handling
- Introduced a new task to update all devices in tasks.json.
- Changed distance input fields in config.html to accept integer values.
- Updated distance management in config.js and data.cpp to use integers instead of floats.
- Enhanced sensor data handling in Sensor.cpp and Sensor.h for better performance.
- Implemented cache invalidation for distance settings in data.cpp.
- Adjusted WebSocket broadcasts in server.cpp to send status updates more efficiently.

+++ data/config.js
@@ -209,8 +209,8 @@ function updateStatusDisplay(status) {
 }
 
 // Distance Management
-let originalMinDistance = 2.0;
-let originalMaxDistance = 100.0;
+let originalMinDistance = 2;
+let originalMaxDistance = 100;
 
 function loadDistanceSettings() {
     fetch("/get_distance_settings")
@@ -233,7 +233,7 @@ function loadDistanceSettings() {
 function updateMinDistanceButton() {
     const input = document.getElementById("minDistanceInput");
     const button = document.getElementById("saveMinDistanceBtn");
-    const currentValue = parseFloat(input.value);
+    const currentValue = parseInt(input.value);
 
     if (isNaN(currentValue) || currentValue === originalMinDistance) {
         button.disabled = true;
@@ -249,7 +249,7 @@ function updateMinDistanceButton() {
 function updateMaxDistanceButton() {
     const input = document.getElementById("maxDistanceInput");
     const button = document.getElementById("saveMaxDistanceBtn");
-    const currentValue = parseFloat(input.value);
+    const currentValue = parseInt(input.value);
 
     if (isNaN(currentValue) || currentValue === originalMaxDistance) {
         button.disabled = true;
@@ -263,10 +263,10 @@ function updateMaxDistanceButton() {
 }
 
 function saveMinDistance() {
-    let minDistance = parseFloat(
+    let minDistance = parseInt(
         document.getElementById("minDistanceInput").value
     );
-    let maxDistance = parseFloat(
+    let maxDistance = parseInt(
         document.getElementById("maxDistanceInput").value
     );
 
@@ -314,10 +314,10 @@ function saveMinDistance() {
 }
 
 function saveMaxDistance() {
-    let minDistance = parseFloat(
+    let minDistance = parseInt(
         document.getElementById("minDistanceInput").value
     );
-    let maxDistance = parseFloat(
+    let maxDistance = parseInt(
         document.getElementById("maxDistanceInput").value
     );
